automatically backed up by learn

+++ db/migrate/01_create_users_table.rb
@@ -3,7 +3,7 @@ class CreateUsersTable < ActiveRecord::Migration[4.2]
     create_table :users do |t|
       t.string :username
       t.string :password
-      t.decimal :balance
+      t.integer :balance
     end
   end    
 end
